+++ cv_personal_page/cv_personal_page/settings/base.py
@@ -118,7 +118,6 @@ STATICFILES_DIRS = (BASE_DIR / "static/",)
 
 LOCALE_PATHS = (BASE_DIR / "locale/",)
 
-MEDIA_URL = "/media/"
 MEDIA_ROOT = BASE_DIR / "media/"
 
 # Default primary key field type