+++ cv_personal_page/cv_personal_page/urls.py
@@ -15,17 +15,16 @@ Including another URLconf
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
 
-from django.conf import settings
 from django.conf.urls.i18n import i18n_patterns
-from django.conf.urls.static import static
 from django.contrib import admin
 from django.urls import include, path
+
+from .views import favicon_view
 
 urlpatterns = i18n_patterns(
     path("admin/", admin.site.urls),
     path("", include("home.urls")),
     path("i18n/", include("django.conf.urls.i18n")),
-)
-
-if settings.DEBUG:
-    urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
+) + [
+    path("favicon.ico", favicon_view, name="favicon"),
+]