+++ cv_personal_page/home/urls.py
@@ -1,7 +1,8 @@
 from django.urls import path
 
-from home.views import home_view
+from home.views import background_view, home_view
 
 urlpatterns = [
     path("", home_view),
+    path("background.jpg", background_view, name="background"),
 ]